requirements.txt
django>=6.0,<7.0
djangorestframework>=3.17
django-cors-headers>=4.9
gunicorn>=23.0
dj-database-url>=2.3
whitenoise>=6.8
psycopg2-binary>=2.9
